+++ EnhanceQoL/Init.lua
@@ -660,6 +660,9 @@ addon.variables.landingPageType = {
 	[30] = { title = DRAGONFLIGHT_LANDING_PAGE_TITLE, checkbox = EXPANSION_NAME9 },
 	[40] = { title = WAR_WITHIN_LANDING_PAGE_TITLE, checkbox = EXPANSION_NAME10 },
 }
+if MIDNIGHT_LANDING_PAGE_TITLE then
+	addon.variables.landingPageType[50] = { title = MIDNIGHT_LANDING_PAGE_TITLE, checkbox = EXPANSION_NAME11 or MIDNIGHT_LANDING_PAGE_TITLE }
+end
 addon.variables.landingPageReverse = {} -- Used for onShow Method of LandingPage
 for id, data in pairs(addon.variables.landingPageType) do
 	addon.variables.landingPageReverse[data.title] = id